CHAD-17069: zigbee-illuminance lazy load sub drivers
