+++ drivers/SmartThings/zigbee-illuminance-sensor/src/aqara/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FINGERPRINTS = {
+  { mfr = "LUMI", model = "lumi.sen_ill.agl01" }
+}
+
+return FINGERPRINTS
