+++ drivers/SmartThings/zigbee-illuminance-sensor/src/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 local ZigbeeDriver = require "st.zigbee"
@@ -26,12 +16,10 @@ local zigbee_illuminance_driver = {
     capabilities.illuminanceMeasurement,
     capabilities.battery
   },
-  sub_drivers = {
-    require("aqara")
-  },
   lifecycle_handlers = {
     doConfigure = do_configure,
   },
+  sub_drivers = require("sub_drivers"),
   health_check = false,
 }
